Complete Rewrite
It's about time.
