+++ src/ed/tilesetEditor.h
@@ -6,11 +6,11 @@
 #include <string>
 #include <string.h>
 
-struct EditorNew;
+struct Editor;
 
 struct TilesetEditor
 {
-    EditorNew* ed; // Main editor.
+    Editor* ed; // Main editor.
     bool open = false; // Window is open.
     std::string name; // Tileset name.
     RenderTexture2D target; // Where to draw the texture.
@@ -29,7 +29,7 @@ struct TilesetEditor
     int currLayer = -1; // Current layer to paint on.
 
     // Create a tileset editor to look at a tileset.
-    TilesetEditor(EditorNew* ed, std::string name, float tileSize);
+    TilesetEditor(Editor* ed, std::string name, float tileSize);
 
     // Draw the displayed tileset image.
     void Draw();
