+++ src/ed/tools/eraser.cpp
@@ -1,7 +1,7 @@
 #include "eraser.h"
-#include "../editorNew.h"
+#include "../editor.h"
 
-void EraserToolActivate(EditorNew* ed, Vector2 pos1, Vector2 pos2)
+void EraserToolActivate(Editor* ed, Vector2 pos1, Vector2 pos2)
 {
     PxMap& layer = ed->map.maps[ed->currentLayer];
     u8 oldTile = layer.GetTile((u16)pos1.x, (u16)pos1.y);
@@ -9,7 +9,7 @@ void EraserToolActivate(EditorNew* ed, Vector2 pos1, Vector2 pos2)
     if (oldTile != 0) ed->undoStack.PushTileErased(ed, ed->currentLayer, oldTile, (u16)pos1.x, (u16)pos1.y);
 }
 
-void EraserToolUpdate(EditorNew* ed)
+void EraserToolUpdate(Editor* ed)
 {
     EraserToolActivate(ed, { (float)ed->GetTileX(), (float)ed->GetTileY() }, { 0, 0 });
 }
