+++ src/ed/tools/zoom.h
@@ -2,7 +2,7 @@
 
 #include "raylib.h"
 
-struct EditorNew;
+struct Editor;
 
-void ZoomToolActivate(EditorNew* ed, Vector2 pos1, Vector2 pos2);
-void ZoomToolUpdate(EditorNew* ed);
+void ZoomToolActivate(Editor* ed, Vector2 pos1, Vector2 pos2);
+void ZoomToolUpdate(Editor* ed);
