+++ src/ed/undoStack.h
@@ -4,7 +4,7 @@
 #include "../types.h"
 #include <deque>
 
-struct EditorNew;
+struct Editor;
 
 enum class UndoActionType {
     TilePlaced,
@@ -52,16 +52,16 @@ public:
     int MaxLen();
     void Reset();
     void PushUndoAction(UndoAction& a);
-    void PushTilePlaced(EditorNew* ed, u8 layer, u8 oldTile, u8 newTile, u16 xPos, u16 yPos);
-    void PushTileErased(EditorNew* ed, u8 layer, u8 oldTile, u16 xPos, u16 yPos);
-    void PushLevelPropertyEdited(EditorNew* ed);
-    void PushEntityPropertyEdited(EditorNew* ed);
-    void PushEntityMoved(EditorNew* ed, int entityIndex, u16 xPos, u16 yPos, u16 newXPos, u16 newYPos);
-    void PushEntityPlaced(EditorNew* ed, int entityIndex, u8 type, u16 xPos, u16 yPos);
-    void PushEntityDeleted(EditorNew* ed, int entityIndex, Entity e);
+    void PushTilePlaced(Editor* ed, u8 layer, u8 oldTile, u8 newTile, u16 xPos, u16 yPos);
+    void PushTileErased(Editor* ed, u8 layer, u8 oldTile, u16 xPos, u16 yPos);
+    void PushLevelPropertyEdited(Editor* ed);
+    void PushEntityPropertyEdited(Editor* ed);
+    void PushEntityMoved(Editor* ed, int entityIndex, u16 xPos, u16 yPos, u16 newXPos, u16 newYPos);
+    void PushEntityPlaced(Editor* ed, int entityIndex, u8 type, u16 xPos, u16 yPos);
+    void PushEntityDeleted(Editor* ed, int entityIndex, Entity e);
     void SetLastChained();
-    void Undo(EditorNew* ed);
-    void Redo(EditorNew* ed);
+    void Undo(Editor* ed);
+    void Redo(Editor* ed);
     bool CanUndo();
     bool CanRedo();
 };
